Be able to list oauth client
